Created privilege and user setup
For Show Com and Show Chair, also added some constants for the breakout
of the people deserving bios - teachers, performers, special guests,
vendors, but that's not fully complete.

Still need to run through all privilege related settings for someone
who is exclusively a show committee member or chair and need to
separate the pages with specialized filters.