More performance updates
Separated the code for conference bigs and acts, it's a messy design,
but it's better than the massively conjoined cases, and helps if we
have more functional modifications.

Redid navigation, and updated the summary pages to work with filters -
conference is everything BUT performances, and vice versa.  Also, added
a "show committee" guidance page.