BidChoice adds flexibility
Upon making another radio button table, I decided to abstract a bit -
there is now a single table - Bid Choice, that is made for questions
and answers with a limited set (yes, no, etc).  Choose questions by
building an array, and set the acceptable answers.  Could use more
abstraction for database loading and static display, but this was a
good start.