First round of edits for Stage Name
The Stage Name should be the defacto display name.  There are now two
new columns to user "Display Name" (always full) and "Stage Name"
(optional).  The Display Name defaults to Stage Name, using First/Last
as a backup.  From here out, if you see the real name when a stage name
is in the user's profile, it's a bug.