All event types added, and a creature comfort
Now it's possible to create and manage Shows, Drop-In classes, and
Master classes using the Special Events creation process.  They'll show
up in their appropriate areas on the information pages and in the
schedule!

Also - a small creature comfort - by default the "teacher" and "email"
on the class bid form will now show up as the submitter.