Fixing 2 errors when you become user
When you use this admin function (most don't have it), you can't submit
a bid, and you get a weird change profile state if you happen to have
an open id.  That's due to changes I made to the code base, and needed
an update to the become user thread.