Data layer for new schedule display
Data layer is now done, based on gbe_run and gbe_event.  It is tested
and demoed through ScheduleDisplayTest.php, which can be run as a web
page URL.  Also made minor updates to Schedule to fix a bug in
Conference display and to make sure that invisible event times are
consistently invisible.