Refactor and display cleanup
Pulled signup check out of Schedule.php so it can be used more widely
later - it should also get used for scheduling gms, and letting people
book act rehearsal.  Now it's in gbe_event, which is the general area
it vaguely relates so, although it doesn't use this class at all.