Schedule management for shows
Added automatic schedule management for performers and crew.
Performers are added or dropped through Shows->table of acts - when
they are booked in a show, their bio shows up and they get scheduled
forcibly for the show and call time.  Tech Crew can volunteer for a
Tech Rehearsal and they will be signed up if they are free for the
rehearsal, call time and show.  But if they are not free, they won't
get booked.  Anyone on Stage Staff (ie, a "gm") can see the list of
signups for what ever event they have been made staff for (tech
rehearsal shows crew, call shows everyone, shows show everyone) - and
the user information is annotated if it's a performer.  Also fixed
minor bugs in schedule display and eliminated some dead code related to
gender and tshirts.  This code is in the center of the spagetti, so the
code changes affect a diverse set of files and it's pretty ugly.