Added Order & viable defaults
ActTechDisplay must be seeded for a given show for the query form to
show up for performers.  The checkin actdisplaydefaults.sql lets you
seed your database with the display data that fits our four actual
shows - but you need to customize for your databse - when using this
form, replace the name of the show with the event id in the events
table for the connected show.