Make it impossible to edit the bid for an accepted act.
There's a way to do it if they try hard enough - and adminstrators
could still do it - I didn't access control it.  But I make it
impossible through the GUI - there is no longer a link anywhere that
lets you edit the bid for something that's accepted.  Also - for
accepted classes and panels, you can now go to the description for your
class/panel and see what it says (instead of the bid) so that you can
edit it if you have the power.