one space!
"Crispus Attucks (Master/Drop-in Classes)" had two spaces which broke
with the database expectations.  Really, in the end, the constant
variables like these should be killed and we should create pull downs
off the enums in the database, or use the constants and leave the
database more open.  This is unnecessary pain.

Pardon the rant.